Unique cron to manage all objects to delete from db.
